some reorganization/renaming
